requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openai==1.3.0
sentence-transformers>=2.5.0
faiss-cpu==1.13.2
numpy>=1.26.0
flask==3.0.0
python-dotenv==1.0.0
pandas>=2.1.0
tqdm==4.66.1
PyPDF2==3.0.1
pdfplumber==0.10.3